requirements.txt
fastapi==0.115.4
supabase==2.3.8
uvicorn==0.32.0
pydantic==2.6.3
pydantic[email]==2.6.3
pyjwt==2.9.0
python-dotenv==1.0.1
httpx==0.25.2
gotrue==2.1.0
tenacity==9.0.0
pybreaker==1.2.0
requests==2.31.0
kafka-python==2.0.2
asyncio==3.4.3
prometheus-fastapi-instrumentator==7.0.0